src/pyproject.toml
[tool.poetry]
name = "sequencer"
version = "0.1.0"
description = ""
authors = ["Wonjae Choi <choi@wonj.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
web3 = "^6.19.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
